// src/app/interfaces/RespApi.ts
export interface Role {
    value: string;
    name:  string;
}

export interface GetStatus {
    status: Status[];
}

export interface Status {
    value: string;
    name:  string;
}


export interface GetSeverities {
    severities: Severity[];
}

export interface Severity {
    value: string;
    name:  string;
}


export interface GetActivateUsers {
    status: boolean;
    users:  ChicoServicio[];
}

export interface ChicoServicio {
    _id:      string;
    name:     string;
    username: string;
    image:    string;
    role:     Role;
    online:   boolean;
    isActive: boolean;
}